Move ConditionType to static member of Condition
