+++ kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/operator/ClusterCondition.java
@@ -18,7 +18,7 @@ import edu.umd.cs.findbugs.annotations.Nullable;
 import static java.util.stream.Collectors.joining;
 
 public record ClusterCondition(@NonNull String cluster,
-                               @NonNull ConditionType type,
+                               @NonNull Condition.Type type,
                                @NonNull Condition.Status status,
                                @Nullable String reason,
                                @Nullable String message) {
@@ -26,21 +26,21 @@ public record ClusterCondition(@NonNull String cluster,
     public static final String INVALID = "Invalid";
 
     static ClusterCondition accepted(String cluster) {
-        return new ClusterCondition(cluster, ConditionType.Accepted, Condition.Status.TRUE, null, null);
+        return new ClusterCondition(cluster, Condition.Type.Accepted, Condition.Status.TRUE, null, null);
     }
 
     static ClusterCondition filterInvalid(String cluster, String filterName, String detail) {
-        return new ClusterCondition(cluster, ConditionType.Accepted, Condition.Status.FALSE, INVALID,
+        return new ClusterCondition(cluster, Condition.Type.Accepted, Condition.Status.FALSE, INVALID,
                 String.format("Filter \"%s\" is invalid: %s", filterName, detail));
     }
 
     public static ClusterCondition refNotFound(String cluster, LocalRef<?> ref) {
-        return new ClusterCondition(cluster, ConditionType.Accepted, Condition.Status.FALSE, INVALID,
+        return new ClusterCondition(cluster, Condition.Type.Accepted, Condition.Status.FALSE, INVALID,
                 String.format("Resource of kind \"%s\" in group \"%s\" named \"%s\" does not exist.", ref.getKind(), ref.getGroup(), ref.getName()));
     }
 
     public static ClusterCondition filterNotFound(String cluster, String filterName) {
-        return new ClusterCondition(cluster, ConditionType.Accepted, Condition.Status.FALSE, INVALID,
+        return new ClusterCondition(cluster, Condition.Type.Accepted, Condition.Status.FALSE, INVALID,
                 String.format("Filter \"%s\" does not exist.", filterName));
     }
 
@@ -48,7 +48,7 @@ public record ClusterCondition(@NonNull String cluster,
         String ingresses = ingressConflictExceptions.stream()
                 .map(IngressConflictException::getIngressName)
                 .collect(joining(","));
-        return new ClusterCondition(cluster, ConditionType.Accepted, Condition.Status.FALSE, INVALID,
+        return new ClusterCondition(cluster, Condition.Type.Accepted, Condition.Status.FALSE, INVALID,
                 String.format("Ingress(es) [%s] of cluster conflicts with another ingress", ingresses));
     }
